requirements.txt
playwright>=1.40.0
pyyaml>=6.0
aiohttp>=3.9.0
flask>=3.0.0
